+++ meinberlin/apps/dashboard/views.py
@@ -1,36 +1,14 @@
-from django.contrib.messages.views import SuccessMessageMixin
 from django.core.urlresolvers import reverse
-from django.utils.translation import ugettext_lazy as _
-from django.views import generic
 
 from meinberlin.apps.dashboard2 import mixins as a4dashboard_mixins
 from meinberlin.apps.dashboard2 import views as a4dashboard_views
 from meinberlin.apps.newsletters.forms import NewsletterForm
 from meinberlin.apps.newsletters.views import NewsletterCreateView
-from meinberlin.apps.organisations.models import Organisation
-
-from . import forms
 
 
 class DashboardProjectListView(a4dashboard_views.ProjectListView):
     def get_queryset(self):
         return super().get_queryset().filter(projectcontainer=None)
-
-
-class DashboardOrganisationUpdateView(a4dashboard_mixins.DashboardBaseMixin,
-                                      SuccessMessageMixin,
-                                      generic.UpdateView):
-
-    model = Organisation
-    form_class = forms.OrganisationForm
-    slug_url_kwarg = 'organisation_slug'
-    template_name = 'meinberlin_dashboard/organisation_form.html'
-    success_message = _('Organisation successfully updated.')
-    permission_required = 'meinberlin_organisations.change_organisation'
-    menu_item = 'organisation'
-
-    def get_permission_object(self):
-        return self.organisation
 
 
 class DashboardNewsletterCreateView(a4dashboard_mixins.DashboardBaseMixin,
